+++ libs/ngrx-toolkit/src/lib/with-storage-sync.spec.ts
@@ -1,10 +1,10 @@
 import { getState, patchState, signalStore, withState } from '@ngrx/signals';
 import { TestBed } from '@angular/core/testing';
 import * as flushPromises from 'flush-promises';
-import { StorageServiceFactory } from '../internal/storage.service';
-import { withIndexeddb } from '../features/with-indexeddb';
-import { withLocalStorage } from '../features/with-local-storage';
-import { withStorageSync } from '../with-storage-sync';
+import { StorageServiceFactory } from './storage-sync/internal/storage.service';
+import { withIndexeddb } from './storage-sync/features/with-indexeddb';
+import { withLocalStorage } from './storage-sync/features/with-local-storage';
+import { withStorageSync } from './with-storage-sync';
 
 interface StateObject {
   foo: string;
